Raise exception when find more than one matched Cinder volumes
For method fuxi.volumeprovider.cinder.Cinder._get_docker_volume,
It just return the first Cinder volume if find with search_opts,
or None if not find.
But for Fuxi only one volume could exist on backend with the same
name and metadata which contains key 'volume_from', so when find
more than one matched Cinder volume, this method should raise an
exception.

Change-Id: I08542a77b15f3938995463e577fdbf4a4e9e39b4
Closes-Bug: #1631925