Raise exception when find more than one matched Cinder volumes
For method fuxi.volumeprovider.cinder.Cinder._get_docker_volume,
It just return the first Cinder volume if find with search_opts,
or None if not find.
But for Fuxi only one volume could exist on backend with the same
name and metadata which contains key 'volume_from', so when find
more than one matched Cinder volume, this method should raise an
exception.

Change-Id: I6b4af9151363a3b6de8294b836f0a46e64566245
Closes-Bug: #1631925