Retrieve volumes from cinder with matched metadata
For fuxi, it should retrieves volumes from Cinder which as long as
have matched metadata(link {"volume_from": "fuxi"}) when Docker
executes cmd  'docker volume ls'.
For condition 'or not vol.attachment', it means that volume must
be attached at least one host, so that Fuxi could recognize it,
so revome this condition.

Change-Id: Ie73ebc301d51d33cbce924121ce6971761177542
Closes-Bug: #1636365