Search volumes in all tenants
If users pre-create cinder volumes in their own tenants and pass
the cinder volume ID to Fuxi, Fuxi will return successfully and
the docker volume was created. However, Fuxi will fail to search
the pre-created volumes if users list/inspect the volume. This
is because Fuxi tried to find volumes in 'service' tenant but
the volume was pre-created in users tenants. This commit
fixed it.

Change-Id: Idb4a3611465aa44cf88aa0abe144897849a27b7d