Update test requirement
Since pbr already landed and the old version of hacking seems not
work very well with pbr>=2, we should update it to match global
requirement.

In addition, this patch disable multiattach in fullstack tests
since this option was disabled in Cinder [1]

NOTE: This commit actually contains two fixes: one is for fixing
the test requirement and the other is for fixing the fullstack
tests. Ideally, we should split it into two commits but we are not
able to do it because we need both fixes in one patch to unblock
the gate.

[1] https://review.openstack.org/#/c/428365/

Closes-Bug: #1669494
Partial-Bug: #1668848
Change-Id: I6770eda1cf232aa8528190c82f1868a82372b159