[WIP]Convert attaching volume from sync to async
For cinder provider with connector type 'openstack', volumes will be attached
to host one by one when creating multi-volume at the same time.
With this way, it will spend much time at waiting for attaching.
In this patch, this situation will be replaced with attaching volume asychronously.
Of course, this way is main for kvm hypervisor.

Change-Id: I4f8f05d240dfb6cc2dbdbe5b8c0c4189cfb4e290
blueprint: attaching-cinder-volume-asychronously