Reuse neutron-lib for pep8 hacking checks
neutron library has a good hacking rules set, we can use these rules
instead of reinvent the wheel. These such rules are:

- [N521] Validate that jsonutils module is used instead of json
- [N532] Validate that LOG.warning is used instead of LOG.warn. The
         latter is deprecated.
- [N537] Don't translate logs.
- [N534] Exception messages should be translated

It also enables H904, which allows the logging package to skip
creating the formatted log message if the message is not going
to be emitted because of the current log level.

Change-Id: I3b3649bbe8a7a3ae69d3ea1553923567b5a350d7
Closes-Bug: #1704691